lint: Break down files
