+++ src/app/[variants]/(main)/resource/library/_layout/Sidebar.tsx
@@ -4,14 +4,14 @@ import { memo } from 'react';
 
 import { NavPanelPortal } from '@/features/NavPanel';
 import SideBarLayout from '@/features/NavPanel/SideBarLayout';
-import FileTree from '@/features/ResourceManager/components/Tree';
+import LibraryHierarchy from '@/features/ResourceManager/components/LibraryHierarchy';
 
 import Header from './Header';
 
 const Sidebar = memo(() => {
   return (
     <NavPanelPortal navKey="resourceLibrary">
-      <SideBarLayout body={<FileTree />} header={<Header />} />
+      <SideBarLayout body={<LibraryHierarchy />} header={<Header />} />
     </NavPanelPortal>
   );
 });
