+++ src/features/ResourceManager/components/LibraryHierarchy/types.ts
@@ -1,0 +1,10 @@
+export interface TreeItem {
+  children?: TreeItem[];
+  fileType: string;
+  id: string;
+  isFolder: boolean;
+  name: string;
+  slug?: string | null;
+  sourceType?: string;
+  url: string;
+}
